Rename cmdlib to libcmd
