+++ gateway/cmd/cmd.go
@@ -1,16 +1,16 @@
 package cmd
 
-import "github.com/ubclaunchpad/pinpoint/cmdlib"
+import "github.com/ubclaunchpad/pinpoint/libcmd"
 
 // GatewayCommand is the CLI for pinpoint-gateway
 type GatewayCommand struct {
-	*cmdlib.Command
+	*libcmd.Command
 }
 
 // New creates a new GatewayCommand
 func New(version string) *GatewayCommand {
 	app := &GatewayCommand{
-		Command: cmdlib.New("pinpoint-gateway",
+		Command: libcmd.New("pinpoint-gateway",
 			"Pinpoint's RESTful API gateway",
 			``,
 			version,
